refactor(api): move ChallengeAlreadyAnsweredError to shared
