+++ api/tests/certification/evaluation/unit/domain/usecases/create-live-alert_test.js
@@ -1,8 +1,8 @@
 import _ from 'lodash';
 import sinon from 'sinon';
 
-import { ChallengeAlreadyAnsweredError } from '../../../../../../src/certification/evaluation/domain/errors.js';
 import { createLiveAlert } from '../../../../../../src/certification/evaluation/domain/usecases/create-live-alert.js';
+import { ChallengeAlreadyAnsweredError } from '../../../../../../src/shared/domain/errors.js';
 import { expect } from '../../../../../test-helper.js';
 import { domainBuilder } from '../../../../../tooling/domain-builder/domain-builder.js';
 import { catchErr } from '../../../../../tooling/test-utils/error.js';
